refactor: remove legacy capture.go
Move shared parsing helpers to artifacts.go and remove unused
capture.go and capture_test.go files.
